requirements.txt
azure-functions
azureml-core
numpy
pandas
scikit-learn
pymsteams
